test(swap): dedupe findSwapQuote suite fixtures into a shared module
Extract the coin pair and minimal quote builders shared by the selection
and ranked suites into __tests__/swapQuoteFixtures so the new ranked test
no longer clones the selection test's setup block (jscpd gate). vi.mock
registrations stay per-file since vitest hoists them per test module.
Also export BoundSwapQuote through the tools and root barrels so consumers
can name SwapQuoteCandidate.quote.
