+++ packages/sdk/src/tools/swap/index.ts
@@ -1,6 +1,12 @@
 export type { AcrossChain, AcrossQuote, AcrossQuoteParams } from './acrossQuote'
 export { acrossQuote, acrossSupportedChains } from './acrossQuote'
-export type { FindSwapQuoteParams, FindSwapQuotesResult, SwapQuote, SwapQuoteCandidate } from './findSwapQuote'
+export type {
+  BoundSwapQuote,
+  FindSwapQuoteParams,
+  FindSwapQuotesResult,
+  SwapQuote,
+  SwapQuoteCandidate,
+} from './findSwapQuote'
 export { findSwapQuote, findSwapQuotes } from './findSwapQuote'
 export type { JupiterQuoteResponse, JupiterSwapParams, JupiterSwapResult } from './jupiter'
 export {
